fix: fixed build error and types
