+++ app/dice-sets/page.tsx
@@ -1,6 +1,5 @@
-import { DiceSetSelector } from "@/src/components/dice-set-selector";
+import { DiceSetSelector } from "@/components/dice-set-selector";
 import { fetchDiceSets } from "@/app/actions";
-import { DiceSet } from "@/src/types/dice";
 
 export default async function DiceSetsPage() {
   const initialDiceSets = await fetchDiceSets();
