+++ components/dice-set-selector.tsx
@@ -2,9 +2,9 @@
 
 import { useState } from "react";
 import { useRouter } from "next/navigation";
-import { Button } from "@/src/components/ui/button";
-import { Input } from "@/src/components/ui/input";
-import { Label } from "@/src/components/ui/label";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
 import { addDiceSet } from "@/app/actions";
 import { useFormStatus } from "react-dom";
 import { DiceSet } from "@/src/types/dice";
