+++ components/hero.tsx
