--- src/components/typography/inline-code.tsx
@@ -1,7 +1,0 @@
-export function TypographyInlineCode() {
-  return (
-    <code className="relative rounded bg-muted px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold">
-      @radix-ui/react-alert-dialog
-    </code>
-  );
-}
